chore: move array files around to make them smaller chunks
