+++ src/index.ts
@@ -1,6 +1,4 @@
-export * from "./Array.ts"
 export * as Array from "./Array.ts"
-export type { AnyArray, NonEmptyArray, ReadonlyNonEmptyArray } from "./Array.ts"
 export * as Either from "./Either.ts"
 export * as Future from "./Future.ts"
 export * as Option from "./Option.ts"
